requirements.txt
Flask==3.0.0
line-bot-sdk==1.8.0
gunicorn==21.2.0
openai>=1.0.0
aiohttp==3.8.6
stripe
psycopg2-binary
psutil==5.9.5
